security 적용

+++ Spring Boot_블로그/src/main/java/blog/blogProject/entity/Board.java
@@ -1,5 +1,6 @@
 package blog.blogProject.entity;
 
+import lombok.Builder;
 import lombok.Data;
 
 import javax.persistence.*;
@@ -11,7 +12,7 @@ public class Board {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String title;
     private String content;
